Implement the init command (#4)
* feat: Configure the file system with a local disk

The disk name uses the app name to create the folder. The disk folder will be created in the current working directory: `./.drzl`

* feat: Include the manifest folder to dockerignore file

* feat: Set to www-data user the UID 1000 in the docker container

This fix the issue:
Create the manifiest files with the right permission. With the default value the root was the file owner.

* feat: Implements a Manifiest class to handle the file

* feat: Defines the Manifest class as singleton

* feat: Implement the InitCommand class

It generates a manifest file with default values. Also, prompt the user to update the manifest file with their specific configuration.

* feat: Implements the create and exists methods

* feat: add the guard method ensureManifestDoesntExist

Also, use the Manifest::create method to initialize the manifest file.

Now, the handle method has a clearer implementation.

* feat: Remove Orchestra dependency

Uses LaravelZero\TestCase class as a BaseTestCase

* feat: Defines the TestCase for the test suite.

* feat: Define a basic Pest.php file

* feat: Configure PHPUnit for Pest

* feat: Import from Laravel Zero the CreateApplication trait

* feat: Ignore the storage folder for git

* style: Add a blank line at the end of the file.

* refactor: Injects the FilesystemAdapter into the Manifest class

It allows a fake adapter to be injected during the test run.

* fix: Fake Storage is not injected into the Manifest

If we include the SummaryCommand the tests fail because the fake storage is not properly injected into the Manifest class.

* test: Implement a basic test for init command

* refactor: Deletes the TestCase::getPackageProviders method

LaravelZero already loads the service provider properly.

* refactor: Deletes TestCase::tearDown method

For the time being it is not necessary.

* refactor: Remove the use dor DrzlServiceProvider

* test: Add expects output for the create manifest file

* refactor: Rename the Manifest::create method to Manifest::created

It improves the readability of the code as it is used.

* refactor: Replace if by match to report the manifest creation

* refactor: Uses match to drive all cases of the handle method

The match is a more explicit and readable construct when we have multiple cases, and may be more suitable for handling boolean conditions or adding new cases in the future.

* refactor: Changes the visibility of Manifest::filename to public

It is necessary to validate if the file exists from the test suite.

* test: Ensure the manifest file is created by the init command

* test: Ensure the manifest file is created with default values

* test: Adds a test to ensure the user is prompted to update the file

* refactor: Update test description

* test: Verifies the manifest file is not overwritten

* refactor: Update test name to begin with a verb